chatPanelView dividing refactoring
