+++ Desktop/MIN.Desktop/Contracts/Models/FileUploadingStatus.cs
@@ -1,0 +1,16 @@
+﻿using MIN.Common.Core.Contracts.Interfaces;
+using MIN.Desktop.Contracts.Interfaces;
+
+namespace MIN.Desktop.Contracts.Models;
+
+/// <summary>
+/// UI Статус отправки файла на сервер
+/// </summary>
+public record struct FileUploadingStatus(Guid TransferId, string FileName, string SenderName, string FileSize) : IDescribableStatus
+{
+    /// <inheritdoc />
+    public readonly Guid Id => TransferId;
+
+    /// <inheritdoc />
+    string IDescribable.GetDescription() => $"{SenderName} отправляет файл {FileName} {FileSize}";
+}
