+++ Desktop/MIN.Desktop/Views/Panels/PanelViews/ChatPanel/ChatPanelView.UI.cs
@@ -1,0 +1,237 @@
+﻿using MIN.Core.Messaging.RoomRelated;
+using MIN.Core.Transport.NamedPipes.Models;
+using MIN.Desktop.Components;
+using MIN.Desktop.Contracts.Constants;
+using MIN.Desktop.Contracts.Interfaces;
+using MIN.Desktop.Contracts.Schemes;
+using MIN.Desktop.Infrastructure.Events;
+using MIN.Helpers.Services;
+
+namespace MIN.Desktop.Views.Panels.PanelViews.ChatPanel;
+
+public partial class ChatPanelView
+{
+    private readonly List<IDescribableStatus> currentStatuses = [];
+
+    /// <inheritdoc />
+    protected override void ApplyStylings()
+    {
+        splitContainer.Panel1.BackColor = ColorScheme.PrimaryAccent;
+        splitContainer.Panel2.BackColor = ColorScheme.MainPanelBackground;
+        splitContainerSideBar.Panel2.BackColor = ColorScheme.PrimaryAccent;
+        tableLayoutPanelHeader.BackColor = ColorScheme.PrimaryAccent;
+        tableLayoutPanelStats.BackColor = ColorScheme.PrimaryAccent;
+        notificationComboBox.BackColor = ColorScheme.PrimaryAccent;
+
+        participantsInfo.ForeColor = ColorScheme.TextOnAccent;
+        hostName.ForeColor = ColorScheme.TextOnAccent;
+        computer.ForeColor = ColorScheme.TextOnAccent;
+        classroom.ForeColor = ColorScheme.TextOnAccent;
+        createdAt.ForeColor = ColorScheme.TextOnAccent;
+        notificationComboBox.ForeColor = ColorScheme.TextOnAccent;
+        Title.ForeColor = ColorScheme.TextOnAccent;
+        statusLabel.ForeColor = ColorScheme.TextPrimary;
+
+        hostNameLabel.ForeColor = ColorScheme.TextOnAccent;
+        classroomLabel.ForeColor = ColorScheme.TextOnAccent;
+        computerLabel.ForeColor = ColorScheme.TextOnAccent;
+        onlineLabel.ForeColor = ColorScheme.TextOnAccent;
+        participantsLabel.ForeColor = ColorScheme.TextOnAccent;
+        createdAtLabel.ForeColor = ColorScheme.TextOnAccent;
+
+        participantsFlow.BackColor = ColorScheme.DividerColor;
+        chatFlow.BackColor = ColorScheme.ChatAreaBackground;
+        chatFlow.Padding = new Padding(chatFlow.Padding.Left, chatFlow.Padding.Top, chatFlow.Padding.Right, messageMinPadding);
+
+        tableLayoutPanelButtons.RowStyles[0] = new RowStyle(SizeType.AutoSize);
+    }
+
+    #region Update
+
+    private void UpdateStats()
+    {
+        if (room == null)
+        {
+            return;
+        }
+
+        Text = $"MIN - Комната {room.Name}";
+        Title.Text = $"Комната {room.Name}";
+
+        var isHost = room.HostParticipant?.Id == localParticipant.Id;
+        hostName.Text = isHost ? "Ты" : room.HostParticipant?.Name ?? "Неизвестно";
+        createdAt.Text = room.CreatedAt.ToShortTimeString();
+        editButton.Visible = isHost;
+
+        if (CollegePCNameParser.TryParseComputerName(endpoint is NamedPipeEndpoint npEndpoint
+                ? npEndpoint.MachineName
+                : string.Empty,
+            out var roomNumber,
+            out var computerNumber))
+        {
+            computer.Text = computerNumber.ToString();
+            classroom.Text = roomNumber.ToString();
+        }
+        else
+        {
+            computer.Text = DesktopConstants.UndefinedPCName;
+            classroom.Text = DesktopConstants.UndefinedPCName;
+        }
+
+        UpdateParticipantFlow();
+    }
+
+    private void UpdateParticipantFlow()
+    {
+        participantsFlow.Controls.Clear();
+
+        if (room == null)
+        {
+            return;
+        }
+
+        foreach (var participant in room.CurrentParticipants)
+        {
+            var card = new ParticipantCard(participant,
+                isHost: participant.Id == room.HostParticipant.Id,
+                isSelf: participant.Id == localParticipant.Id)
+            {
+                Width = participantsFlow.Width - participantsFlow.Margin.Horizontal * 2,
+            };
+
+            card.OnCardContextMenuStripClicked += (selected, particpant) =>
+            {
+                privateChatParticipantId = selected ? participant.Id : null;
+            };
+
+            participantsFlow.Controls.Add(card);
+        }
+
+        participantsInfo.Text = $"{room.ParticipantCount}/{room.MaximumParticipants}";
+    }
+
+    private void UpdateChatFlow()
+    {
+        chatFlow.Controls.Clear();
+
+        if (room == null)
+        {
+            return;
+        }
+
+        foreach (var storedMessage in room.ChatHistory)
+        {
+            AddMessageToChatFlow(storedMessage);
+        }
+    }
+
+    #endregion
+
+    #region Helper methods
+
+    private void SendSystemMessage(SystemTextMessage systemMessage, bool needsToNotify = false)
+    {
+        AddMessageToChatFlow(systemMessage);
+
+        if (needsToNotify)
+        {
+            featureCollection.Core.EventBus.PublishAsync(new DescribableMessageReceivedEvent()
+            {
+                RoomId = roomId,
+                DescribableMessage = systemMessage,
+            });
+            NotifyIfNeeded(systemMessage);
+        }
+    }
+
+    private void AddStatus(IDescribableStatus status)
+    {
+        currentStatuses.Add(status);
+        ShowStatusRow();
+        statusLabel.Text = string.Join(", ", currentStatuses.Select(x => x.GetDescription()));
+    }
+
+    private void RemoveStatus(Guid statusId)
+    {
+        var foundStatus = currentStatuses.FirstOrDefault(x => x.Id == statusId);
+        if (foundStatus != null)
+        {
+            currentStatuses.Remove(foundStatus);
+        }
+
+        if (currentStatuses.Count == 0)
+        {
+            HideStatusRow();
+        }
+        else
+        {
+            statusLabel.Text = string.Join(", ", currentStatuses.Select(x => x.GetDescription()));
+        }
+    }
+
+    #endregion
+
+    #region Extra rows
+
+    private void HideMultiFileAttachmentUploader(bool withClear = false)
+    {
+        if (withClear)
+        {
+            multiFileAttachmentUploader.Clear();
+        }
+
+        var row = tableLayoutPanelButtons.GetRow(multiFileAttachmentUploader);
+
+        tableLayoutPanelButtons.SuspendLayout();
+
+        tableLayoutPanelButtons.RowStyles[row].Height = 0;
+
+        multiFileAttachmentUploader.Visible = false;
+        multiFileAttachmentUploader.OnLastFileRemoved = null;
+
+        tableLayoutPanelButtons.ResumeLayout(true);
+        ResizeMessageTextBox();
+    }
+
+    private void ShowMultiFileAttachmentUploader()
+    {
+        var row = tableLayoutPanelButtons.GetRow(multiFileAttachmentUploader);
+
+        tableLayoutPanelButtons.SuspendLayout();
+
+        tableLayoutPanelButtons.RowStyles[row].Height = 76;
+
+        multiFileAttachmentUploader.Visible = true;
+
+        tableLayoutPanelButtons.ResumeLayout(true);
+        ResizeMessageTextBox();
+    }
+
+    private void HideStatusRow()
+    {
+        var row = tableLayoutPanelButtons.GetRow(statusLabel);
+
+        tableLayoutPanelButtons.SuspendLayout();
+
+        tableLayoutPanelButtons.RowStyles[row].Height = 0;
+        statusLabel.Visible = false;
+
+        tableLayoutPanelButtons.ResumeLayout(true);
+        ResizeMessageTextBox();
+    }
+
+    private void ShowStatusRow()
+    {
+        var row = tableLayoutPanelButtons.GetRow(statusLabel);
+
+        tableLayoutPanelButtons.SuspendLayout();
+
+        tableLayoutPanelButtons.RowStyles[row].Height = 15;
+        statusLabel.Visible = true;
+
+        tableLayoutPanelButtons.ResumeLayout(true);
+        ResizeMessageTextBox();
+    }
+
+    #endregion
+}
