+++ Desktop/MIN.Desktop/Views/Panels/PanelViews/ChatPanel/ChatPanelView.resx
