+++ ISSUETODOLIST.txt
@@ -1,3 +1,2 @@
 1. Сделать превью изображений
 2. В каком-то месте неплохо было бы увидеть кнопку "Сохранить как.."
-3. Сделать статус об выгрузке файле от клиента к хосту
